+++ Simple Game Library/CHANGES.md
@@ -6,6 +6,7 @@ Date: *2016-04-??*
 - Added method `getCurrentContext()` to retrieve the current graphics context
 in situations when `beginUpdate()` has already been called and the returned context
 is not available.
+- Added Vector2d class.
 
 ## Version 1.1.0
 Date: *2016-04-05*